webui: move export naming and download back to the conversation store
