+++ tools/server/webui/src/lib/components/app/chat/ChatSettings/ChatSettingsImportExportTab.svelte
@@ -2,7 +2,7 @@
 	import { Download, Upload, Trash2 } from '@lucide/svelte';
 	import { Button } from '$lib/components/ui/button';
 	import { DialogConversationSelection, DialogConfirmation } from '$lib/components/app';
-	import { createMessageCountMap, downloadConversationFile } from '$lib/utils';
+	import { createMessageCountMap } from '$lib/utils';
 	import { ISO_DATE_TIME_SEPARATOR } from '$lib/constants';
 	import { conversationsStore, conversations } from '$lib/stores/conversations.svelte';
 	import { toast } from 'svelte-sonner';
@@ -56,7 +56,7 @@
 				})
 			);
 
-			downloadConversationFile(
+			conversationsStore.downloadConversationFile(
 				allData,
 				`${new Date().toISOString().split(ISO_DATE_TIME_SEPARATOR)[0]}_conversations.json`
 			);
