+++ tools/server/webui/src/lib/utils/conversation-utils.ts
@@ -1,19 +1,7 @@
 /**
  * Utility functions for conversation data manipulation
  */
-import type { ExportedConversations, DatabaseMessage } from '$lib/types';
-import {
-	ISO_DATE_TIME_SEPARATOR,
-	ISO_DATE_TIME_SEPARATOR_REPLACEMENT,
-	ISO_TIMESTAMP_SLICE_LENGTH,
-	EXPORT_CONV_ID_TRIM_LENGTH,
-	EXPORT_CONV_NONALNUM_REPLACEMENT,
-	EXPORT_CONV_NAME_SUFFIX_MAX_LENGTH,
-	ISO_TIME_SEPARATOR,
-	ISO_TIME_SEPARATOR_REPLACEMENT,
-	NON_ALPHANUMERIC_REGEX,
-	MULTIPLE_UNDERSCORE_REGEX
-} from '$lib/constants';
+import type { DatabaseMessage } from '$lib/types';
 
 /**
  * Creates a map of conversation IDs to their message counts from exported conversation data
@@ -40,63 +28,4 @@ export function createMessageCountMap(
  */
 export function getMessageCount(conversationId: string, countMap: Map<string, number>): number {
 	return countMap.get(conversationId) ?? 0;
-}
-
-/**
- * Generates a sanitized filename for a conversation export
- * @param conversation - The conversation metadata
- * @param msgs - Optional array of messages belonging to the conversation
- * @returns The generated filename string
- */
-export function generateConversationFilename(
-	conversation: { id?: string; name?: string },
-	msgs?: DatabaseMessage[]
-): string {
-	const conversationName = (conversation.name ?? '').trim().toLowerCase();
-
-	const sanitizedName = conversationName
-		.replace(NON_ALPHANUMERIC_REGEX, EXPORT_CONV_NONALNUM_REPLACEMENT)
-		.replace(MULTIPLE_UNDERSCORE_REGEX, '_')
-		.substring(0, EXPORT_CONV_NAME_SUFFIX_MAX_LENGTH);
-
-	// If we have messages, use the timestamp of the newest message
-	const referenceDate = msgs?.length
-		? new Date(Math.max(...msgs.map((m) => m.timestamp)))
-		: new Date();
-
-	const iso = referenceDate.toISOString().slice(0, ISO_TIMESTAMP_SLICE_LENGTH);
-	const formattedDate = iso
-		.replace(ISO_DATE_TIME_SEPARATOR, ISO_DATE_TIME_SEPARATOR_REPLACEMENT)
-		.replaceAll(ISO_TIME_SEPARATOR, ISO_TIME_SEPARATOR_REPLACEMENT);
-	const trimmedConvId = conversation.id?.slice(0, EXPORT_CONV_ID_TRIM_LENGTH) ?? '';
-	return `${formattedDate}_conv_${trimmedConvId}_${sanitizedName}.json`;
-}
-
-/**
- * Triggers a browser download of the provided exported conversation data
- * @param data - The exported conversation payload (either a single conversation or array of them)
- * @param filename - Filename; if omitted, a deterministic name is generated
- */
-export function downloadConversationFile(data: ExportedConversations, filename?: string): void {
-	// Choose the first conversation or message
-	const conversation = 'conv' in data ? data.conv : Array.isArray(data) ? data[0]?.conv : undefined;
-	const msgs =
-		'messages' in data ? data.messages : Array.isArray(data) ? data[0]?.messages : undefined;
-
-	if (!conversation) {
-		console.error('Invalid data: missing conversation');
-		return;
-	}
-
-	const downloadFilename = filename ?? generateConversationFilename(conversation, msgs);
-
-	const blob = new Blob([JSON.stringify(data, null, 2)], { type: 'application/json' });
-	const url = URL.createObjectURL(blob);
-	const a = document.createElement('a');
-	a.href = url;
-	a.download = downloadFilename;
-	document.body.appendChild(a);
-	a.click();
-	document.body.removeChild(a);
-	URL.revokeObjectURL(url);
 }
