+++ tools/server/webui/src/lib/utils/index.ts
@@ -41,12 +41,7 @@ export { setConfigValue, getConfigValue, configToParameterRecord } from './confi
 export { buildProxiedUrl, getProxiedUrlString } from './cors-proxy';
 
 // Conversation utilities
-export {
-	createMessageCountMap,
-	getMessageCount,
-	generateConversationFilename,
-	downloadConversationFile
-} from './conversation-utils';
+export { createMessageCountMap, getMessageCount } from './conversation-utils';
 
 // Clipboard utilities
 export {
